Move FaceCornerIterator to DomainHelpers
